Move LookupLemmatizer to spacy.lemmatizer
